+++ docs/.vuepress/configs/navbar.ts
@@ -1,0 +1,32 @@
+import type { NavbarConfig } from '@vuepress/theme-default';
+
+
+/**
+ * Navbar links.
+ */
+export const navbar: NavbarConfig = [
+    {
+        text: "Tutorials",
+        link: "/tutorial/"
+    },
+    {
+        text: "Reference",
+        link: "/reference/"
+    },
+    {
+        text: "Publications",
+        link: "/publication/"
+    },
+    {
+        text: "Developer",
+        link: "/developer/"
+    },
+    {
+        text: "News",
+        link: "https://github.com/mihaiconstantin/powerly/blob/main/NEWS.md"
+    },
+    {
+        text: "CRAN",
+        link: "https://CRAN.R-project.org/package=powerly"
+    }
+]