+++ NEWS.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## development
+### Added
+- Add starter package documentation via `VuePress`. The deployed documentation
+  files can be accessed at [powerly.dev](https://powerly.dev).
+
 ## 1.8.2 - 2022-07-22
 ### Fixed
 - Remove square brackets from version numbers in `NEWS.md` to allow CRAN to